Add beaker tests
Add basic tests to ensure the puppet manifests can be applied cleanly
and the service will start.

We could test with a real IRC network and a real registered IRC user,
but the service starts just fine even if it can't connect to an IRC
server, and it's not the puppet module's job to test the bot itself,
it's sufficient to test that the server is configured the way we expect
it to be.

Change-Id: If153cdbfcd0a78b25ca01f22a10ab1c9055c0766